Hide equivalence class codes as an implementation detail
initially, the idea was to be able to define custom equivalence classes
from user code. However, realistically, nobody wants to do this, so we
should not complicate things here and just follow our usual pattern.
