+++ libflatsurf/src/combinatorial_equivalence.cc
@@ -23,9 +23,9 @@
 
 #include "impl/combinatorial_equivalence.hpp"
 
-#include "../flatsurf/equivalence_class_code.hpp"
 #include "../flatsurf/edge.hpp"
 #include "../flatsurf/iterable.hpp"
+#include "impl/equivalence_class_code.hpp"
 #include "util/assert.ipp"
 #include "util/hash.ipp"
 
@@ -33,88 +33,6 @@ namespace flatsurf {
 
 template <typename Surface>
 CombinatorialEquivalence<Surface>::CombinatorialEquivalence(Predicate predicate): predicate(predicate) {}
-
-template <typename Surface>
-std::unique_ptr<EquivalenceClassCode> CombinatorialEquivalence<Surface>::code(const Surface& surface) const {
-  // We describe the process for combinatorial equivalence; the general
-  // algorithm is described below.
-
-  // Pick a random half edge as the starting point and call it half edge 1.
-  // Then walk the surface in a breadth-first search, labeling half edges as 2,
-  // 3, ... as we find unlabeled ones. The sequence of half edges we see during
-  // this walk is a code for this surface. If we sort the codes and pick the
-  // lexicographically largest one, then it uniquely describes this surface
-  // combinatorially.
-
-  // To speed things up we compute all the codes in parallel and discard codes
-  // that are not the lexicographically largest immediately.
-  // As an optimization we don't need to record the label of the edge we are
-  // crossing since it just going to be smallest edge label we have seen so
-  // far.
-
-  // Note that things get a bit more complicated here because we support that
-  // there is a predicate on the half edges that singles out possibly
-  // non-triangular cells that should be the only relevant data for this notion
-  // of equivalence.
-
-  // The procedure for generic codes is similar. We start with some initial
-  // "walkers" that emit a code word in each of their steps. We compare these
-  // code words and discard any code word that is not minimal und all the
-  // minimal walkers are completed. This yields the minimal code for this surface.
-
-  Code code;
-
-  using Walker = CombinatorialWalker;
-
-  std::vector<Walker> walkers = Walker::seedWalkers(surface, predicate);
-
-  std::vector<Walker*> minimalWalkers;
-
-  for (auto& walker: walkers)
-    minimalWalkers.push_back(&walker);
-
-  while (true) {
-    std::vector<Walker*> stillMinimalWalkers;
-
-    std::optional<Word> minimalWord;
-
-    for (auto* walker : minimalWalkers) {
-      const auto word = walker->walk();
-
-      if (!word.has_value())
-        continue;
-
-      switch (Walker::cmp(word, minimalWord)) {
-        case -1:
-          minimalWord = word;
-          stillMinimalWalkers = {walker};
-          break;
-        case 0:
-          stillMinimalWalkers.push_back(walker);
-          break;
-        case 1:
-          // We drop this walker. It does not lead to a minimal code (unless all
-          // walkers are finished.)
-          break;
-        default:
-          LIBFLATSURF_UNREACHABLE("cmp() must return -1, 0, or 1");
-      }
-
-    }
-
-    LIBFLATSURF_ASSERT(stillMinimalWalkers.empty() == !minimalWord.has_value(), "when there is a minimal walker it must have produced a new code word and vice versa");
-
-    if (!minimalWord.has_value())
-      // All walkers are done. Our code is complete.
-      break;
-
-    Walker::append(code, *minimalWord);
-
-    minimalWalkers = stillMinimalWalkers;
-  }
-
-  return std::make_unique<CombinatorialEquivalenceClassCode>(code);
-}
 
 template <typename Surface>
 Iterable<Deformation<Surface>> CombinatorialEquivalence<Surface>::automorphisms() const {
@@ -185,6 +103,25 @@ std::optional<typename CombinatorialEquivalence<Surface>::Word> CombinatorialEqu
 }
 
 template <typename Surface>
+std::unique_ptr<EquivalenceClassCode> CombinatorialEquivalence<Surface>::code(const Surface& surface) const {
+  return ImplementationOf<Equivalence<Surface>>::code(seedWalkers(surface));
+}
+
+template <typename Surface>
+std::vector<typename CombinatorialEquivalence<Surface>::CombinatorialWalker> CombinatorialEquivalence<Surface>::seedWalkers(const Surface& surface) const {
+  std::vector<CombinatorialWalker> walkers;
+
+  for (const auto start : surface.halfEdges()) {
+    if (!predicate(surface, start))
+      continue;
+
+    walkers.push_back({surface, start, predicate});
+  }
+
+  return walkers;
+}
+
+template <typename Surface>
 int CombinatorialEquivalence<Surface>::CombinatorialWalker::cmp(const std::optional<Word>& lhs, const std::optional<Word>& rhs) {
   if (!lhs.has_value()) {
     if (!rhs.has_value())
@@ -236,20 +173,6 @@ int CombinatorialEquivalence<Surface>::CombinatorialWalker::label(const Surface&
   labeled.push_back(halfEdge);
 
   return label;
-}
-
-template <typename Surface>
-std::vector<typename CombinatorialEquivalence<Surface>::CombinatorialWalker> CombinatorialEquivalence<Surface>::CombinatorialWalker::seedWalkers(const Surface& surface, const Predicate& predicate) {
-  std::vector<CombinatorialWalker> walkers;
-
-  for (const auto start : surface.halfEdges()) {
-    if (!predicate(surface, start))
-      continue;
-
-    walkers.push_back({surface, start, predicate});
-  }
-
-  return walkers;
 }
 
 template <typename Surface>
