+++ libflatsurf/src/equivalence.cc
@@ -24,15 +24,13 @@
 #include "../flatsurf/edge.hpp"
 #include "impl/combinatorial_equivalence.hpp"
 #include "impl/linear_equivalence.hpp"
+#include "util/assert.ipp"
 
 namespace flatsurf {
 
 template <typename Surface>
-Equivalence<Surface>::Equivalence(std::shared_ptr<ImplementationOf<Equivalence>> impl) : self(impl) {}
-
-template <typename Surface>
 Equivalence<Surface> Equivalence<Surface>::combinatorial(std::function<bool(const Surface&, Edge)> predicate) {
-  return Equivalence(std::make_unique<CombinatorialEquivalence<Surface>>(predicate));
+  return Equivalence(PrivateConstructor{}, std::make_shared<CombinatorialEquivalence<Surface>>(predicate));
 }
 
 template <typename Surface>
@@ -56,7 +54,7 @@ Equivalence<Surface> Equivalence<Surface>::areaPreserving(bool oriented, std::fu
 
 template <typename Surface>
 Equivalence<Surface> Equivalence<Surface>::linear(bool oriented, std::function<Matrix(const Surface&, HalfEdge, HalfEdge)> normalization, std::function<bool(const Surface&, Edge)> predicate) {
-  return Equivalence(std::make_unique<LinearEquivalence<Surface>>(oriented, normalization, predicate));
+  return Equivalence(PrivateConstructor{}, std::make_shared<LinearEquivalence<Surface>>(oriented, normalization, predicate));
 }
 
 template <typename Surface>
@@ -76,6 +74,88 @@ typename Equivalence<Surface>::Matrix Equivalence<Surface>::orthonormalization(c
 
 template <typename Surface>
 ImplementationOf<Equivalence<Surface>>::~ImplementationOf() {}
+
+template <typename Surface>
+template <typename Walker>
+std::unique_ptr<EquivalenceClassCode> ImplementationOf<Equivalence<Surface>>::code(const std::vector<Walker>& walkers) const {
+  // We describe the process for combinatorial equivalence; the general
+  // algorithm is described below.
+
+  // Pick a random half edge as the starting point and call it half edge 1.
+  // Then walk the surface in a breadth-first search, labeling half edges as 2,
+  // 3, ... as we find unlabeled ones. The sequence of half edges we see during
+  // this walk is a code for this surface. If we sort the codes and pick the
+  // lexicographically largest one, then it uniquely describes this surface
+  // combinatorially.
+
+  // To speed things up we compute all the codes in parallel and discard codes
+  // that are not the lexicographically largest immediately.
+  // As an optimization we don't need to record the label of the edge we are
+  // crossing since it just going to be smallest edge label we have seen so
+  // far.
+
+  // Note that things get a bit more complicated here because we support that
+  // there is a predicate on the half edges that singles out possibly
+  // non-triangular cells that should be the only relevant data for this notion
+  // of equivalence.
+
+  // The procedure for generic codes is similar. We start with some initial
+  // "walkers" that emit a code word in each of their steps. We compare these
+  // code words and discard any code word that is not minimal und all the
+  // minimal walkers are completed. This yields the minimal code for this surface.
+
+  using Code = typename Walker::Code;
+  using Word = typename Walker::Word;
+
+  Code code;
+
+  std::vector<Walker*> minimalWalkers;
+
+  for (auto& walker: walkers)
+    minimalWalkers.push_back(&walker);
+
+  while (true) {
+    std::vector<Walker*> stillMinimalWalkers;
+
+    std::optional<Word> minimalWord;
+
+    for (auto* walker : minimalWalkers) {
+      const auto word = walker->walk();
+
+      if (!word.has_value())
+        continue;
+
+      switch (Walker::cmp(word, minimalWord)) {
+        case -1:
+          minimalWord = word;
+          stillMinimalWalkers = {walker};
+          break;
+        case 0:
+          stillMinimalWalkers.push_back(walker);
+          break;
+        case 1:
+          // We drop this walker. It does not lead to a minimal code (unless all
+          // walkers are finished.)
+          break;
+        default:
+          LIBFLATSURF_UNREACHABLE("cmp() must return -1, 0, or 1");
+      }
+
+    }
+
+    LIBFLATSURF_ASSERT(stillMinimalWalkers.empty() == !minimalWord.has_value(), "when there is a minimal walker it must have produced a new code word and vice versa");
+
+    if (!minimalWord.has_value())
+      // All walkers are done. Our code is complete.
+      break;
+
+    Walker::append(code, *minimalWord);
+
+    minimalWalkers = stillMinimalWalkers;
+  }
+
+  return Walker::code(code);
+}
 
 template <typename Surface>
 std::ostream& operator<<(std::ostream&, const Equivalence<Surface>& surface) {
